Update pSlip.py
Patch Note: Proper Handling of <activity-alias> Exported Status

Previously pSlip would flag an <activity-alias> as vulnerable if only the alias was exported  even when its underlying <activity> was not exported. This could result in false positives—activities that are actually blocked at runtime were still being flagged as externally launchable.

With this patch, the scanner now checks both the alias’s android:exported and the real <activity>’s android:exported. If the real <activity> is not exported, the alias is no longer flagged. This ensures pSlip only report vulnerabilities for <activity-alias> entries that can truly be reached from outside the app.